requirements.txt
flask
playwright
gunicorn
python-dotenv
Flask-Limiter
redis

